Book index and show pages render
Moved BookListing into Book module
