+++ assets/js/app.js
@@ -4,7 +4,6 @@ import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 import axios from "axios";
 import Book from "./components/book";
-import BookListing from "./components/bookListing";
 import Books from "./components/books"
 import Home from "./components/home"
 import LogIn from "./components/logIn"
