+++ assets/js/components/book.js
@@ -3,7 +3,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 import axios from "axios";
-import BookListing from "./bookListing";
 import Books from "./books"
 import Home from "./home"
 import LogIn from "./logIn"
