+++ assets/js/components/home.js
@@ -3,7 +3,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
 import Book from "./book";
-import BookListing from "./bookListing";
 import Books from "./books"
 import LogIn from "./logIn"
 import NewBook from "./newBook"
