+++ assets/js/components/signUp.js
@@ -4,7 +4,6 @@ import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 import axios from "axios";
 import Book from "./book";
-import BookListing from "./bookListing";
 import Books from "./books"
 import Home from "./home"
 import LogIn from "./logIn"
